Create custom hook for login
