+++ SopPro/app/(auth)/test.jsx
@@ -1,12 +1,15 @@
-import { View, Text } from 'react-native'
-import React from 'react'
+import { View, Text } from "react-native";
+import React from "react";
+import { SafeAreaView } from "react-native-safe-area-context";
 
 const test = () => {
   return (
     <View>
-      <Text>test logged in test</Text>
+      <SafeAreaView>
+        <Text>Test - second protected route</Text>
+      </SafeAreaView>
     </View>
-  )
-}
+  );
+};
 
-export default test
+export default test;
